split category reducer
